# openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Zoning API
  description: |-
    An API for serving data related to the zoning of New York City.
  version: 0.1.0
servers:
  - url: https://zoning.planningdigital.com/api 
...
paths:
...
  /zoning-districts/{id}:
    get:
      summary: Non-spatial details for a specific zoning district
      operationId: findZoningDistrictByZoningDistrictId
      tags:
      - Zoning Districts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A zoning district object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ZoningDistrict'
        '400':
          $ref: "#/components/responses/BadRequest"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /zoning-districts/{id}/classes:
    get:
      summary: Class schemas for the specified zoning district
      operationId: findZoningDistrictClassesByZoningDistrictId
      tags:
      - Zoning Districts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: An object of class schemas for the zoning district.
          content:
            application/json:
              schema:
                type: object
                properties:
                  zoningDistrictClasses:
                    type: array
                    items:
                      $ref: '#/components/schemas/ZoningDistrictClass'
                required:
                  - zoningDistrictClasses
        '400':
          $ref: "#/components/responses/BadRequest"
        '404':
          $ref: "#/components/responses/NotFound"
        '500':
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  schemas:
...
    ZoningDistrict:
      type: object
      properties:
        id:
          type: string
          description: An automatically generated uuid.
          format: uuid
          example: d1c09f3a-553b-4574-811e-abd59d19e01b
        label:
          type: string
          description: The zoning codes that apply to the district. Multiple codes are concatenated with a slash.
          example: M1-5/R7-3
      required:
        - id
        - label
    ZoningDistrictClass:
      type: object
      properties:
        id:
          type: string
          description: The code associated with the Zoning class.
          pattern:  ^[A-Z][0-9]+$
          example: M1
        category:
          $ref: "#/components/schemas/ZoningDistrictClassCategory"
        description:
          type: string
          description: Zoning class descriptions.
        url:
          type: string
          description: Planning website page that explains the Zoning District
        color:
          type: string
          description: Zoning classes from layer groups.
          pattern: '^#([A-Fa-f0-9]{8})$'
          example: '#f3b3ffff'
      required:
        - id
        - category
        - description
        - url
        - color
    ZoningDistrictClassCategory:
      type: string
      enum:
        - "Residential"
        - "Commercial"
        - "Manufacturing"
      description: The type of zoning district.
...
    Error:
      type: object
      properties:
        statusCode:
          type: number
        message:
          type: string
        error:
          type: string
      required:
        - statusCode
        - message
...
  responses:
...
    BadRequest:
      description: Invalid client request
      content:
        application/json:
          schema: 
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Server side error
      content:
        application/json:
          schema: 
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Requested resource does not exist or is not available
      content:
        application/json:
          schema: 
            $ref: "#/components/schemas/Error"